SoftwareDesignPatternsProject
In this project I will demonstrate my knowledge of design patterns, as
well as an ability to evaluate the pros and cons of using, or not using,
them in various design situations. As part of this I will also
demonstrate a knowledge of design principles, code refactoring, and code
smells.